# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Physical Venue Risk Assessment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessments/{id}/risk-matrix:
    get:
      operationId: getRiskMatrix
      tags: [riskMatrix]
      summary: Get risk matrix for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk matrix
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskMatrix"
        "404":
          description: Not found
    put:
      operationId: upsertRiskMatrix
      tags: [riskMatrix]
      summary: Create or update risk matrix for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskMatrixInput"
      responses:
        "200":
          description: Risk matrix saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskMatrix"
...
  /assessments/{id}/evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidence"
    post:
      operationId: createEvidence
      tags: [evidence]
      summary: Add evidence to an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
  /evidence/{id}:
    patch:
      operationId: updateEvidence
      tags: [evidence]
      summary: Update evidence item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
    delete:
      operationId: deleteEvidence
      tags: [evidence]
      summary: Delete evidence item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create an incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    delete:
      operationId: deleteIncident
      tags: [incidents]
      summary: Delete an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /venues/{id}/osint:
    get:
      operationId: getVenueOsint
      tags: [osint]
      summary: Get OSINT suggestions for a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OSINT events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OsintEvent"
  /osint/{id}/review:
    patch:
      operationId: reviewOsintEvent
      tags: [osint]
      summary: Accept or reject an OSINT finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OsintReviewInput"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OsintEvent"
components:
  schemas:
...
    RiskMatrix:
      type: object
      required: [id, assessmentId, areaRisk, accessControl, arrivalDeparture, parking, personnel, medical, hse, extraction, overallRating]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        areaRisk:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        accessControl:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        arrivalDeparture:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        parking:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        personnel:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        medical:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        hse:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        extraction:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        overallRating:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    RiskMatrixInput:
      type: object
      properties:
        areaRisk:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        accessControl:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        arrivalDeparture:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        parking:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        personnel:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        medical:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        hse:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        extraction:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        overallRating:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        notes:
          type: string
...
    Incident:
      type: object
      required: [id, incidentType, severity, incidentDate, summary, verified, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: ["integer", "null"]
        venueName:
          type: ["string", "null"]
        incidentType:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        incidentDate:
          type: string
        summary:
          type: string
        sourceName:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        distanceFromVenue:
          type: ["number", "null"]
        confidenceLevel:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
    IncidentInput:
      type: object
      required: [incidentType, severity, incidentDate, summary]
      properties:
        venueId:
          type: integer
        incidentType:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        incidentDate:
          type: string
        summary:
          type: string
        sourceName:
          type: string
        sourceUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        distanceFromVenue:
          type: number
        confidenceLevel:
          type: string
          enum: [low, medium, high]
        verified:
          type: boolean
    IncidentUpdate:
      type: object
      properties:
        incidentType:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        incidentDate:
          type: string
        summary:
          type: string
        sourceName:
          type: string
        sourceUrl:
          type: string
        verified:
          type: boolean
    Evidence:
      type: object
      required: [id, assessmentId, evidenceType, label, verified, createdAt]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        evidenceType:
          type: string
        label:
          type: string
        content:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        filename:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        analystNote:
          type: ["string", "null"]
        verified:
          type: boolean
        uploadedByName:
          type: ["string", "null"]
        createdAt:
          type: string
    EvidenceInput:
      type: object
      required: [evidenceType, label]
      properties:
        evidenceType:
          type: string
          enum: [osint_source, analyst_note, image, video, pdf, website, news_article, police_advisory, government_bulletin, map_screenshot]
        label:
          type: string
        content:
          type: string
        url:
          type: string
        filename:
          type: string
        section:
          type: string
        analystNote:
          type: string
        uploadedBy:
          type: integer
    EvidenceUpdate:
      type: object
      properties:
        label:
          type: string
        content:
          type: string
        url:
          type: string
        section:
          type: string
        analystNote:
          type: string
        verified:
          type: boolean
...
    OsintEvent:
      type: object
      required: [id, venueId, eventType, summary, status, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        eventType:
          type: string
        summary:
          type: string
        sourceName:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        analystNote:
          type: ["string", "null"]
        createdAt:
          type: string
    OsintReviewInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected]
        analystNote:
          type: string